% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlay.r
\name{overlay}
\alias{overlay}
\title{Match Point Geometries to Polygon Geometries (e.g., Voter Locations to School Districts)}
\usage{
overlay(
  points = NULL,
  polygons = NULL,
  point_id = "point_id",
  polygon_id = "polygon_id",
  used_NCES = FALSE,
  FIPS_code = NULL,
  FIPS_col = NULL
)
}
\arguments{
\item{points}{sf object containing point geometries for units (e.g., voter locations).}

\item{polygons}{sf object containing polygon geometries (e.g., school districts).}

\item{point_id}{Name of the column in the points sf object that contains unique identifiers for each point (default: "point_id"). Preferably as string.}

\item{polygon_id}{Name of the column in the polygons sf object that contains unique identifiers for each polygon (default: "polygon_id"). Preferably as string.}

\item{used_NCES}{Boolean indicating whether the user input NCES school district shapefiles or other shapefiles with a state FIPS_code as the polygons (default: FALSE).}

\item{FIPS_code}{State FIPS code to filter NCES or other school district shapefiles (default: NULL, which means no filtering by state).}

\item{FIPS_col}{Column name in your dataset containing FIPS_code. Preferably should be a character/string variable (default: NULL, which means no filtering by state).}
}
\value{
A tibble with three columns: point_id (string), polygon_id (string), and distance (numeric, meters to internal point), where each point_id is matched to one polygon_id.
}
\description{
The \code{overlay()} function performs the "Overlay" (Step 2) operation from the QOR Method.
}
\details{
This function:
\itemize{
\item Matches point geometries (e.g., voter locations) to polygon geometries (e.g., school districts) using spatial intersections and distances.
\item Requires that all points and polygons each have a unique identifier column.
\item If observations are not unique (e.g., panel data), use only one timepoint per function call (e.g., all voters in one year, then all voters in the next year, etc...).
\item Returns a match that assigns \strong{one} polygon to each point: the polygon the point is in, or the closest polygon (based on internal point) if the point is not in any polygon.
\item To return multiple polygons per point, you will need to modify the code.
}
}
