% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.r
\name{query}
\alias{query}
\title{Geocode Unit (Voter) Addresses to Long/Lat Coordinates}
\usage{
query(
  units = NULL,
  street = "street",
  city = "city",
  state = "state",
  state_shape = NULL,
  units_per_batch = 4000,
  year = NULL
)
}
\arguments{
\item{units}{dataframe object containing voter information (needs to have unique ID and then street, city, and state columns)}

\item{street}{name of the column in the units dataframe that contains the street address (default: "street")}

\item{city}{name of the column in the units dataframe that contains the city (default: "city")}

\item{state}{name of the column in the units dataframe that contains the state (default: "state")}

\item{state_shape}{sf object containing the shape of the state (used to filter geocoding results to the state, catching errors)}

\item{units_per_batch}{number of units to geocode in each batch (conservative default: 4000)}

\item{year}{of the data (default: NULL, which means no filtering by year)}
}
\value{
list object with two items: 1. tibble of matched units with their geocoded coordinates, and 2. tibble of unmatched units (those that could not be geocoded).
}
\description{
The query() function performs the "Query" (Step 1) operation from the QOR Method.
This function assigns units (e.g., voters) longitude and latitude coordinates based on
their street address, city, and state. It is, functionally, a wrapper around tidygeocoder::geocode()
that anticipates actions needed to process individual voter data and set up the next QOR Method steps.
Importantly, it does not perform any matching to polygons or zip codes; it simply geocodes the addresses.
}
