% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover.r
\name{recover}
\alias{recover}
\title{Recover Failed Point (e.g., Voter) to Polygon (e.g., School District) Matches using Zipcodes}
\usage{
recover(
  units = NULL,
  polygons = NULL,
  zipcodes = NULL,
  unit_id = "unit_id",
  unit_zip = "postalcode",
  polygon_id = "polygon_id",
  zip_id = "postalcode"
)
}
\arguments{
\item{units}{dataframe object containing voter information (needs to have unique ID and then a zipcode variable)}

\item{polygons}{sf object containing polygon geometries (e.g., school districts)}

\item{zipcodes}{sf object containing zipcode tabulation area geometries (ZCTAs)}

\item{unit_id}{name of the column in the units dataframe that contains the unique identifiers for each unit (default: "unit_id")}

\item{unit_zip}{name of the column in the units dataframe that contains the zipcodes (default: "postalcode")}

\item{polygon_id}{name of the column in the polygons sf object that contains the unique identifiers for each polygon (default: "polygon_id")}

\item{zip_id}{name of the column in the zipcodes sf object that contains the unique identifiers for each zipcode (default: "postalcode")}
}
\value{
tibble dataset with five columns: unit_id, polygon_id, postalcode, distance, and a binary flag for matched_byzip, where each unit_id is matched to ONE polygon_id.
}
\description{
The recover() function performs the "Recover" (Step 3) operation from the QOR Method.
This function assigns units (e.g., voters) to polygon geometries (e.g., school districts) using zipcodes as spatial cross-walks.
It requires that all units, polygons, and zipcodes each have a unique identifier column. If observations are not unique, then this function may not work as intended.
E.g., if you have panel data, you should use ONE time unit per functional call (e.g., all voters in one year, then all voters in the next year, etc.).
Importantly, the function returns matches that assign ONE polygon to each unit, which is the polygon whose internal point is closest to the center of that unit's zipcode.
The "Recover" operation is intended to be used for cases where the "Query" operation fails to convert an address to a point geometry.
Modify the code accordingly if you want to return multiple polygons for each unit.
}
