% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover.r
\name{recover}
\alias{recover}
\title{Recover Failed Point-to-Polygon Matches Using Zipcodes}
\usage{
recover(
  units = NULL,
  polygons = NULL,
  zipcodes = NULL,
  unit_id = "unit_id",
  unit_zip = "postalcode",
  polygon_id = "polygon_id",
  zip_id = "postalcode",
  state_shape = NULL,
  used_NCES = FALSE,
  FIPS_code = NULL,
  FIPS_col = NULL
)
}
\arguments{
\item{units}{Dataframe or tibble containing unit information (must have unique unit_id and a zipcode variable).}

\item{polygons}{sf object containing polygon geometries (e.g., school districts).}

\item{zipcodes}{sf object containing zipcode tabulation area geometries (ZCTAs).}

\item{unit_id}{Name of the column in the units dataframe that contains unique identifiers for each unit (default: "unit_id"). Preferably as string.}

\item{unit_zip}{Name of the column in the units dataframe that contains the zipcodes (default: "postalcode").}

\item{polygon_id}{Name of the column in the polygons sf object that contains unique identifiers for each polygon (default: "polygon_id"). Preferably as string.}

\item{zip_id}{Name of the column in the zipcodes sf object that contains unique identifiers for each zipcode (default: "postalcode"). Preferably as string.}

\item{state_shape}{sf object containing the shape of the state (used to filter zipcodes to the state).}

\item{used_NCES}{Boolean indicating whether the user input NCES school district shapefiles or other shapefiles with a state FIPS_code as the polygons (default: FALSE).}

\item{FIPS_code}{State FIPS code to filter NCES school district shapefiles (default: NULL, which means no filtering by state).}

\item{FIPS_col}{Column name in your polygons dataset containing FIPS_code. Preferably should be a character/string variable (default: NULL, which means no filtering by state).}
}
\value{
A tibble with five columns: unit_id (string), polygon_id (string), postalcode (string), distance (numeric, meters), and a binary flag for matched_byzip, where each unit_id is matched to one polygon_id.
}
\description{
The \code{recover()} function performs the "Recover" (Step 3) operation from the QOR Method.
}
\details{
This function:
\itemize{
\item Assigns units (e.g., voters) to polygon geometries (e.g., school districts) using zipcodes as a spatial crosswalk.
\item Requires that all units, polygons, and zipcodes each have a unique identifier column.
\item If observations are not unique (e.g., panel data), use only one timepoint per function call (e.g., all voters in one year, then all voters from the next year, etc.).
\item Returns a match that assigns \strong{one} polygon to each unit: the polygon whose internal point is closest to the centroid of that unit's zipcode.
\item Intended for cases where the "Query" operation fails to convert an address to a point geometry.
\item To return multiple polygons per unit, you will need to modify the code.
}
}
